+++ mayan/apps/source_staging_folders/tests/test_migrations.py
@@ -2,7 +2,7 @@ from mayan.apps.testing.tests.base import MayanMigratorTestCase
 
 
 class SourceBackendPathMigrationTestCase(MayanMigratorTestCase):
-    migrate_from = ('sources', '0029_update_source_backend_paths')
+    migrate_from = ('sources', '0028_auto_20210905_0558')
     migrate_to = (
         'source_staging_folders', '0001_update_source_backend_paths'
     )
@@ -22,7 +22,7 @@ class SourceBackendPathMigrationTestCase(MayanMigratorTestCase):
             app_label='sources', model_name='Source'
         )
 
-        self.assertTrue(
+        self.assertEqual(
             Source.objects.get(label='test source staging folder').backend_path,
             'mayan.apps.source_staging_folders.source_backends.staging_folder_backends.SourceBackendStagingFolder'
         )