+++ mayan/apps/source_watch_folders/tests/test_migrations.py
@@ -2,7 +2,7 @@ from mayan.apps.testing.tests.base import MayanMigratorTestCase
 
 
 class SourceBackendPathMigrationTestCase(MayanMigratorTestCase):
-    migrate_from = ('sources', '0029_update_source_backend_paths')
+    migrate_from = ('sources', '0028_auto_20210905_0558')
     migrate_to = ('source_watch_folders', '0001_update_source_backend_paths')
 
     def prepare(self):
@@ -11,7 +11,7 @@ class SourceBackendPathMigrationTestCase(MayanMigratorTestCase):
         )
 
         Source.objects.create(
-            backend_path='mayan.apps.sources.source_backends.source_watch_folders.SourceBackendWatchFolder',
+            backend_path='mayan.apps.sources.source_backends.watch_folder_backends.SourceBackendWatchFolder',
             label='test source watch folder'
         )
 
@@ -20,7 +20,7 @@ class SourceBackendPathMigrationTestCase(MayanMigratorTestCase):
             app_label='sources', model_name='Source'
         )
 
-        self.assertTrue(
+        self.assertEqual(
             Source.objects.get(label='test source watch folder').backend_path,
             'mayan.apps.source_watch_folders.source_backends.SourceBackendWatchFolder'
         )