+++ mayan/apps/source_web_forms/tests/test_migrations.py
@@ -2,7 +2,7 @@ from mayan.apps.testing.tests.base import MayanMigratorTestCase
 
 
 class SourceBackendPathMigrationTestCase(MayanMigratorTestCase):
-    migrate_from = ('sources', '0029_update_source_backend_paths')
+    migrate_from = ('sources', '0028_auto_20210905_0558')
     migrate_to = ('source_web_forms', '0001_update_source_backend_paths')
 
     def prepare(self):
@@ -20,7 +20,7 @@ class SourceBackendPathMigrationTestCase(MayanMigratorTestCase):
             app_label='sources', model_name='Source'
         )
 
-        self.assertTrue(
+        self.assertEqual(
             Source.objects.get(label='test source web form').backend_path,
             'mayan.apps.source_web_forms.source_backends.SourceBackendWebForm'
         )